% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbClearResult,SFSQLResult-method}
\alias{dbClearResult,SFSQLResult-method}
\title{dbClearResult}
\usage{
\S4method{dbClearResult}{SFSQLResult}(res, ...)
}
\arguments{
\item{res}{SFSQLResult}

\item{...}{ignored}
}
\description{
dbClearResult for SFSQLResult
}
