% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbDataType,SFSQLDriver,ANY-method}
\alias{dbDataType,SFSQLDriver,ANY-method}
\title{dbDataType}
\usage{
\S4method{dbDataType}{SFSQLDriver,ANY}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{SFSQLConnnection}

\item{obj}{object}

\item{...}{ignored}
}
\description{
dbDataType for SFSQLDriver
}
