% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbDisconnect,SFSQLConnection-method}
\alias{dbDisconnect,SFSQLConnection-method}
\title{dbDisconnect}
\usage{
\S4method{dbDisconnect}{SFSQLConnection}(conn, ...)
}
\arguments{
\item{conn}{SFSQLConnection}

\item{...}{ignored}
}
\description{
dbDisconnect for SFSQLConnection
}
