% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbExistsTable,SFSQLConnection,ANY-method}
\alias{dbExistsTable,SFSQLConnection,ANY-method}
\title{dbExistsTable}
\usage{
\S4method{dbExistsTable}{SFSQLConnection,ANY}(conn, name, ...)
}
\arguments{
\item{conn}{SFSQLConnnection}

\item{name}{table name ('layer' in GDAL)}

\item{...}{arguments to generic}
}
\description{
dbExistsTable for SFSQLConnection
}
