% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbFetch,SFSQLResult-method}
\alias{dbFetch,SFSQLResult-method}
\title{dbFetch}
\usage{
\S4method{dbFetch}{SFSQLResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \code{DBIResult}, created by \code{dbSendQuery()}}

\item{n}{maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Retrieve records from SFSQL query
}
