% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbGetInfo,SFSQLDriver-method}
\alias{dbGetInfo,SFSQLDriver-method}
\title{dbGetInfo}
\usage{
\S4method{dbGetInfo}{SFSQLDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{SFSQLDriver}

\item{...}{ignored}
}
\description{
dbGetInfo for SFSQLDriver
}
