% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbHasCompleted,SFSQLResult-method}
\alias{dbHasCompleted,SFSQLResult-method}
\title{dbHasCompleted}
\usage{
\S4method{dbHasCompleted}{SFSQLResult}(res, ...)
}
\arguments{
\item{res}{SFSQLResult}

\item{...}{ignored}
}
\description{
dbHasCompleted for SFSQLResult
}
\details{
This doesn't do anything
}
