% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbListTables,SFSQLConnection-method}
\alias{dbListTables,SFSQLConnection-method}
\title{dbListTables}
\usage{
\S4method{dbListTables}{SFSQLConnection}(conn, ...)
}
\arguments{
\item{conn}{SFSQLConnnection}

\item{...}{arguments to \code{\link[sf:st_layers]{sf::st_layers()}}}
}
\description{
dbListTables for SFSQLConnection
}
