% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbReadTable,SFSQLConnection,character-method}
\alias{dbReadTable,SFSQLConnection,character-method}
\title{dbReadTable}
\usage{
\S4method{dbReadTable}{SFSQLConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{SFSQLConnnection}

\item{name}{table name ('layer' in GDAL)}

\item{...}{ignored}
}
\description{
dbReadTable for SFSQLConnection
}
\details{
sends a query and fetches the result
}
