% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{GDALSQL}
\alias{GDALSQL}
\title{GDALSQL}
\usage{
GDALSQL()
}
\description{
GDALSQL driver, use with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} to open a data source readable by GDAL
}
\examples{
GDALSQL()
}
\seealso{
lazysf dbConnect
}
