% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALVectorConnection.R, R/connect.R
\docType{class}
\name{GDALVectorConnection-class}
\alias{GDALVectorConnection-class}
\alias{show,GDALVectorConnection-method}
\alias{dbIsValid,GDALVectorConnection-method}
\alias{dbGetInfo,GDALVectorConnection-method}
\alias{dbSendQuery,GDALVectorConnection-method}
\alias{dbReadTable,GDALVectorConnection,character-method}
\alias{dbListTables,GDALVectorConnection-method}
\alias{dbListFields,GDALVectorConnection,character-method}
\alias{dbExistsTable,GDALVectorConnection,ANY-method}
\alias{dbDisconnect,GDALVectorConnection-method}
\title{Class GDALVectorConnection (and methods)}
\usage{
\S4method{show}{GDALVectorConnection}(object)

\S4method{dbIsValid}{GDALVectorConnection}(dbObj, ...)

\S4method{dbGetInfo}{GDALVectorConnection}(dbObj, ...)

\S4method{dbSendQuery}{GDALVectorConnection}(conn, statement, ...)

\S4method{dbReadTable}{GDALVectorConnection,character}(conn, name, ...)

\S4method{dbListTables}{GDALVectorConnection}(conn, ...)

\S4method{dbListFields}{GDALVectorConnection,character}(conn, name, ...)

\S4method{dbExistsTable}{GDALVectorConnection,ANY}(conn, name, ...)

\S4method{dbDisconnect}{GDALVectorConnection}(conn, ...)
}
\description{
GDALVectorConnection objects are created by passing \code{\link[=GDALSQL]{GDALSQL()}} as first
argument to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
They are a superclass of the \link[DBI:DBIConnection-class]{DBI::DBIConnection} class.
The "Usage" section lists the class methods overridden by \pkg{lazysf}.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}, \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}},
\code{\link[DBI:dbReadTable]{DBI::dbReadTable()}}, \code{\link[DBI:dbListFields]{DBI::dbListFields()}},
\code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}, \code{\link[DBI:dbListTables]{DBI::dbListTables()}}.
}
\keyword{internal}
