% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALVectorResult.R
\docType{class}
\name{GDALVectorResult-class}
\alias{GDALVectorResult-class}
\alias{show,GDALVectorResult-method}
\alias{dbFetch,GDALVectorResult-method}
\alias{dbClearResult,GDALVectorResult-method}
\alias{dbHasCompleted,GDALVectorResult-method}
\title{Class GDALVectorResult (and methods)}
\usage{
\S4method{show}{GDALVectorResult}(object)

\S4method{dbFetch}{GDALVectorResult}(res, n = -1, ...)

\S4method{dbClearResult}{GDALVectorResult}(res, ...)

\S4method{dbHasCompleted}{GDALVectorResult}(res, ...)
}
\description{
GDALVectorResult objects are created by \code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}},
and encapsulate the result of an SQL statement.
They are a superclass of the \link[DBI:DBIResult-class]{DBI::DBIResult} class.
The "Usage" section lists the class methods overridden by \pkg{lazysf}.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbFetch]{DBI::dbFetch()}}, \code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}, and
\code{\link[DBI:dbHasCompleted]{DBI::dbHasCompleted()}}.
}
\keyword{internal}
