% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysf-package.R
\docType{package}
\name{lazysf-package}
\alias{lazysf-package}
\title{lazysf: Delayed Read for 'GDAL' Vector Data Sources}
\description{
Lazy read for drawings. A 'dplyr' back end for data sources supported by 'GDAL' vector drivers, that allows working with local or remote sources as if they are in-memory data frames. Basic features work with any drawing format ('GDAL vector data source') supported by the 'gdalraster' package.
}
\section{Package Options}{

There is a debug option \code{options(lazysf.query.debug = TRUE)} which if set
will cause the generated SQL statement to be printed before every call.
In addition it will print the number of rows actually read.

Default geometry format can be set with \code{options(lazysf.geom_format = "WKB")}.
Default SQL dialect can be set with \code{options(lazysf.dialect = "SQLITE")}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/lazysf}
  \item \url{https://hypertidy.github.io/lazysf/}
  \item Report bugs at \url{https://github.com/hypertidy/lazysf/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Authors:
\itemize{
  \item Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})
}

}
\keyword{internal}
