% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{SFSQL}
\alias{SFSQL}
\alias{SFSQL_PG}
\title{SFSQL}
\usage{
SFSQL()

SFSQL_PG()
}
\description{
SFSQL driver
https://gdal.org/user/ogr_sql_dialect.html

SFSQL_PG driver, a wrapper for SFSQL to construct the full connection string required by
GDAL from host,dbname,user,password
}
