% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{SFSQL_PG}
\alias{SFSQL_PG}
\title{SFSQL_PG}
\usage{
SFSQL_PG()
}
\description{
SFSQL_PG driver, a wrapper for SFSQL to construct the full connection string required by
GDAL from host,dbname,user,password
}
