% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL_PGDriver.R
\docType{class}
\name{SFSQL_PGDriver-class}
\alias{SFSQL_PGDriver-class}
\alias{dbDataType,SFSQL_PGDriver,ANY-method}
\alias{dbIsValid,SFSQL_PGDriver-method}
\alias{dbUnloadDriver,SFSQL_PGDriver-method}
\alias{dbGetInfo,SFSQL_PGDriver-method}
\title{Class SFSQL_PGDriver.}
\usage{
\S4method{dbDataType}{SFSQL_PGDriver,ANY}(dbObj, obj, ...)

\S4method{dbIsValid}{SFSQL_PGDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SFSQL_PGDriver}(drv, ...)

\S4method{dbGetInfo}{SFSQL_PGDriver}(dbObj, ...)
}
\description{
SFSQL_PGDriver objects are created by \code{\link[=SFSQL_PG]{SFSQL_PG()}} and used to select the correct
method in \code{\link[=dbConnect]{dbConnect()}}.
They are a superclass of the \linkS4class{DBIDriver} class, and used purely for dispatch.
}
\details{
The "Usage" section lists the class methods overridden by \pkg{lazysf}.
}
\keyword{internal}
