% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQLConnection.R, R/connect.R
\docType{class}
\name{SFSQLConnection-class}
\alias{SFSQLConnection-class}
\alias{show,SFSQLConnection-method}
\alias{dbSendQuery,SFSQLConnection-method}
\alias{dbReadTable,SFSQLConnection,character-method}
\alias{dbListTables,SFSQLConnection-method}
\alias{dbExistsTable,SFSQLConnection,ANY-method}
\alias{dbDisconnect,SFSQLConnection-method}
\title{Class SFSQLConnection (and methods)}
\usage{
\S4method{show}{SFSQLConnection}(object)

\S4method{dbSendQuery}{SFSQLConnection}(conn, statement, ...)

\S4method{dbReadTable}{SFSQLConnection,character}(conn, name, ...)

\S4method{dbListTables}{SFSQLConnection}(conn, ...)

\S4method{dbExistsTable}{SFSQLConnection,ANY}(conn, name, ...)

\S4method{dbDisconnect}{SFSQLConnection}(conn, ...)
}
\description{
SFSQLConnection objects are created by passing \code{\link[=SFSQL]{SFSQL()}} as first
argument to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
They are a superclass of the \link[DBI:DBIConnection-class]{DBI::DBIConnection} class.
The "Usage" section lists the class methods overridden by \pkg{lazysf}.
}
\seealso{
The corresponding generic functions
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}, \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}},
\code{\link[DBI:dbReadTable]{DBI::dbReadTable()}},
\code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}, \code{\link[DBI:dbListTables]{DBI::dbListTables()}}.
}
\keyword{internal}
