% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysf.R
\docType{data}
\name{collect}
\alias{collect}
\title{Force computation of a GDAL query}
\format{
An object of class \code{function} of length 1.
}
\usage{
collect
}
\arguments{
\item{x}{output of \code{\link[=lazysf]{lazysf()}}}

\item{...}{passed to \code{\link[=collect]{collect()}}}
}
\value{
a data frame from \code{collect()}, sf data frame from \code{st_as_sf()}
(only if it contains geometry)
}
\description{
Convert lazysf to an in memory data frame or sf object
}
\details{
\code{collect()} retrieves data into a local table, preserving grouping and ordering.

\code{st_as_sf()} retrieves data into a local sf data frame. Requires the sf
package to be installed, and will succeed only if the result contains a
geometry column (WKB raw vectors). The method is registered when sf is loaded.
}
\examples{
f <- system.file("extdata/multi.gpkg", package = "lazysf", mustWork = TRUE)
lsf <- lazysf(f)
dplyr::collect(lsf)

}
\seealso{
lazysf
}
\keyword{datasets}
