% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFSQL.R
\name{dbConnect,SFSQL_PGDriver-method}
\alias{dbConnect,SFSQL_PGDriver-method}
\title{dbConnect}
\usage{
\S4method{dbConnect}{SFSQL_PGDriver}(
  drv,
  host = "",
  dbname = "",
  user = "",
  password = "",
  readonly = TRUE,
  ...
)
}
\arguments{
\item{drv}{SFSQL_PGDriver created by \code{SFSQL_PG()}}

\item{host}{database server}

\item{dbname}{database name}

\item{user}{user name if needed}

\item{password}{password if needed}

\item{readonly}{open in readonly mode?}

\item{...}{unused}
}
\description{
dbConnect for PostgreSQL via GDAL
}
\details{
https://gdal.org/drivers/vector/pg.html
}
\examples{
\dontrun{
db <- dbConnect(SFSQL_PG(), afile)
dbSendQuery(db,
}
}
