% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALVectorDriver.R
\docType{class}
\name{GDALVectorDriver-class}
\alias{GDALVectorDriver-class}
\alias{dbDataType,GDALVectorDriver,ANY-method}
\alias{dbIsValid,GDALVectorDriver-method}
\alias{dbUnloadDriver,GDALVectorDriver-method}
\alias{dbGetInfo,GDALVectorDriver-method}
\title{Class GDALVectorDriver}
\usage{
\S4method{dbDataType}{GDALVectorDriver,ANY}(dbObj, obj, ...)

\S4method{dbIsValid}{GDALVectorDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{GDALVectorDriver}(drv, ...)

\S4method{dbGetInfo}{GDALVectorDriver}(dbObj, ...)
}
\description{
GDALVectorDriver objects are created by \code{\link[=GDALSQL]{GDALSQL()}} and used to select the correct
method in \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
They are a superclass of the \link[DBI:DBIDriver-class]{DBI::DBIDriver} class, and used purely for dispatch.
}
\details{
The "Usage" section lists the class methods overridden by \pkg{lazysf}.
The \code{\link[DBI:dbDriver]{DBI::dbUnloadDriver()}} method is a null-op.
}
\keyword{internal}
