% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{SFSQL}
\alias{SFSQL}
\title{SFSQL}
\usage{
SFSQL()
}
\description{
SFSQL driver, use to \code{\link[=dbConnect]{dbConnect()}} to a data source readable by sf
}
\examples{
SFSQL()
}
\seealso{
lazysf dbConnect
}
