% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysf.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.tbl_SFSQLConnection}
\title{Convert to in-memory sf}
\usage{
\method{st_as_sf}{tbl_SFSQLConnection}(x, ...)
}
\arguments{
\item{x}{output of \code{\link[=lazysf]{lazysf()}}}

\item{...}{passed to \link{collect}}
}
\value{
a sf data frame (if possible, i.e. contains an \code{sfc} column on \code{\link[=collect]{collect()}})
}
\description{
Convert lazysf to an sf object
}
\details{
WIP: do we need the st_as_sf.data.frame arguments?
}
\examples{
f <- system.file("gpkg/nc.gpkg", package = "sf", mustWork = TRUE)
lsf <- lazysf(f) \%>\% dplyr::select(AREA, FIPS, geom) \%>\% dplyr::filter(AREA < 0.1)
st_as_sf(lsf)
}
